backend/requirements.txt
fastapi>=0.68.0,<1.0.0
uvicorn>=0.15.0,<1.0.0
sqlalchemy>=1.4.23,<3.0.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
python-jose>=3.3.0,<4.0.0
passlib>=1.7.4,<2.0.0
python-multipart>=0.0.5,<1.0.0
alembic>=1.7.4,<2.0.0
psycopg2-binary>=2.9.1,<3.0.0
python-dotenv>=0.19.0,<2.0.0
bcrypt>=3.2.0,<4.0.0
openai>=0.27.0,<3.0.0
email-validator>=2.0.0,<3.0.0